Fixes issue with scheduling a golem/fallen/portal event that hasn't been scheduled before.
Now you can !set portal|fallen|golem even if it hasn't been scheduled before. This used to generate an error because it could not find the event in the scheduled events list.
This also fixes the issue where the specified alliance wasn't being used when returning the newly scheduled event. (!set $alliance_tag $event would return the event for the main alliance instead of $alliance_tag)